+++ packages/calendar/src/components/week-grid/time-grid-event-utils.tsx
@@ -1,13 +1,12 @@
-import { CalendarEventInternal } from '@schedule-x/shared/src/interfaces/calendar/calendar-event.interface'
 import { CustomComponentFn } from '@schedule-x/shared/src/interfaces/calendar/calendar-config'
+import { randomStringId } from '@schedule-x/shared/src/utils/stateless/strings/random'
 
 export const getCCID = (
   customComponent: CustomComponentFn | undefined,
-  calendarEvent: CalendarEventInternal,
   isCopy: boolean | undefined
 ) => {
   let customComponentId = customComponent
-    ? 'custom-time-grid-event-' + calendarEvent.id
+    ? 'custom-time-grid-event-' + randomStringId() // needs a unique string to support event recurrence
     : undefined
   if (customComponentId && isCopy)
     customComponentId += '-' + 'copy' + '-' + calendarEvent.start