+++ packages/event-recurrence/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.37.0](https://github.com/schedule-x/schedule-x/compare/v1.36.0...v1.37.0) (2024-05-01)
+
+**Note:** Version bump only for package @schedule-x/event-recurrence
+
+# [1.36.0](https://github.com/schedule-x/schedule-x/compare/v1.35.0...v1.36.0) (2024-04-30)
+
+**Note:** Version bump only for package @schedule-x/event-recurrence
+
 # [1.35.0](https://github.com/schedule-x/schedule-x/compare/v1.34.0...v1.35.0) (2024-04-26)
 
 **Note:** Version bump only for package @schedule-x/event-recurrence